+++ .env.example
@@ -1,3 +1,8 @@
 DEV_SERVER_PORT=3000
 BUILD_ANALYZE_PORT=8888
 GROUP_NAME=chosica
+DB_NAME=db-name
+DB_USER=db-user
+DB_PASSWORD=db-password
+DB_HOST=127.0.0.1
+DB_PORT=5432
