+++ package.json
@@ -37,6 +37,8 @@
     "express": "^4.18.1",
     "identity-obj-proxy": "^3.0.0",
     "jest-environment-jsdom": "^28.1.2",
+    "pg": "^8.7.3",
+    "pg-hstore": "^2.3.4",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-redux": "^8.0.2",
